requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
autopep8==2.0.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
blurhash==1.1.4
blurhash-python==1.1.3
celery==5.2.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.6
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
kombu==5.2.4
numpy==1.24.2
Pillow==9.4.0
prompt-toolkit==3.0.36
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7.1
redis==4.5.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
vine==5.0.0
wcwidth==0.2.6
